fix: Movido submenu resta a clase Resta como metodo static
